Make common FrameTimer class in the Highbyte.DotNet6502.Systems library that is used by Avalonia, Blazor  WASM and Headless apps.
